add MULTITHREADING

+++ Task5/src/main/java/edu/epam/task5/util/IdGenerator.java
@@ -1,9 +1,9 @@
 package edu.epam.task5.util;
 
 public class IdGenerator {
-    private static long counter = 0;
+    private static int counter = 0;
 
-    public static long generateId() {
+    public static int generateId() {
         return ++counter;
 
     }
